Split and move ApidaeParser to `common` app
Split `geotrek.tourism.ApidaeParser` into:
- `geotrek.common.ApidaeParser` and
- `geotrek.tourism.ApidaeTourismParser`
so the `trekking` app does not depend on the `common` app.
